backend/requirements.txt
# backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
numpy==1.21.2
Pillow==9.3.0
torch==2.2.0
transformers==4.37.2
huggingface-hub==0.20.3
python-jose==3.3.0
python-dotenv==0.19.0